Move constants function to module core::math::constants
