+++ src/core/math/constants.rs
@@ -1,0 +1,3 @@
+pub const INFINITY: f64 = std::f64::INFINITY;
+
+pub const PI: f64 = std::f64::consts::PI;
