+++ src/core/math/mod.rs
@@ -1,2 +1,3 @@
+pub mod constants;
 pub mod optic;
 pub mod rand;
