+++ src/util.rs
@@ -1,5 +1,4 @@
-pub const INFINITY: f64 = std::f64::INFINITY;
-pub const PI: f64 = std::f64::consts::PI;
+use crate::core::math::constants::PI;
 
 pub fn degrees_to_radians(deg: f64) -> f64 {
     deg * PI / 180.0
